requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.3
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
palettable==3.1.1
Pillow==6.0.0
pytz==2019.1
six==1.12.0
Werkzeug==0.15.4
wordcloud==1.5.0
